feat(web): 코스 검색 화면에서 검색 결과가 없는 경우 결과 없음 화면이 나오도록 구현
#61
